+++ src/app/modules/projects/project-settings/project-settings.component.spec.ts
@@ -1,5 +1,9 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+
+import { ThemeService } from '../../../services/theme/theme.service';
+import { NotificationsService } from 'angular2-notifications';
+
 import { ProjectSettingsComponent } from './project-settings.component';
 
 describe('ProjectSettingsComponent', () => {
@@ -9,9 +13,13 @@ describe('ProjectSettingsComponent', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       imports: [BrowserAnimationsModule],
-      declarations: [ ProjectSettingsComponent ]
+      declarations: [ProjectSettingsComponent],
+      providers: [
+        ThemeService,
+        NotificationsService
+      ]
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {