+++ src/app/modules/test-executions/test-executions.component.spec.ts
@@ -1,5 +1,9 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+
+import { ThemeService } from '../../services/theme/theme.service';
+import { NotificationsService } from 'angular2-notifications';
+
 import { TestExecutionsComponent } from './test-executions.component';
 
 describe('TestExecutionsComponent', () => {
@@ -11,7 +15,11 @@ describe('TestExecutionsComponent', () => {
       imports: [
         BrowserAnimationsModule
       ],
-      declarations: [TestExecutionsComponent]
+      declarations: [TestExecutionsComponent],
+      providers: [
+        NotificationsService,
+        ThemeService
+      ]
     })
       .compileComponents();
   }));