+++ src/app/modules/testcases/testcases.component.spec.ts
@@ -3,8 +3,12 @@ import { BaseRequestOptions, Http, HttpModule, RequestMethod, Response, Response
 import { MockBackend, MockConnection } from '@angular/http/testing';
 import { Routes, RouterModule } from '@angular/router';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { AuthenticationService } from '../../services/auth/authentication.service';
+import { FormsModule } from '@angular/forms';
+
 import { ThemeService } from '../../services/theme/theme.service';
+import { NotificationsService } from 'angular2-notifications';
+
+import { AuthenticationService } from '../../services/auth/authentication.service';
 
 import { TestcasesComponent } from './testcases.component';
 import { TestcaseService } from '../../services/testcase/testcase.service';
@@ -21,9 +25,14 @@ describe('TestcasesComponent', () => {
   beforeEach(async(() => {
 
     TestBed.configureTestingModule({
-      imports: [RouterModule, BrowserAnimationsModule],
+      imports: [
+        RouterModule,
+        BrowserAnimationsModule,
+        FormsModule
+      ],
       declarations: [TestcasesComponent, TestsuitesTreeComponent],
       providers: [
+        NotificationsService,
         ThemeService,
         Http,
         AuthenticationService,